requirements.txt
alpha_vantage==3.0.0
fastapi==0.115.8
joblib==1.4.2
numpy==2.2.3
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
scikit_learn==1.6.1
tensorflow==2.18.0
tensorflow_intel==2.18.0
uvicorn==0.34.0
yfinance==0.2.54
